# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hump Yard Market Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CountrySearchInput {
  /** Country name to search */
  country: string;
}

export interface SourceLink {
  url: string;
  title: string;
  /** @nullable */
  publishedDate?: string | null;
  /** @nullable */
  snippet?: string | null;
}

export type KeyContactConfidence = typeof KeyContactConfidence[keyof typeof KeyContactConfidence];


export const KeyContactConfidence = {
  'Named_&_verified': 'Named & verified',
  'Role_known,_name_uncertain': 'Role known, name uncertain',
  Role_inferred: 'Role inferred',
} as const;

export interface KeyContact {
  /**
     * Full name if found, null if only role is known
     * @nullable
     */
  name?: string | null;
  /** Role or job title */
  title: string;
  /** Railway authority or ministry */
  organisation: string;
  /** One-line explanation of why this person matters for BD */
  whyRelevant: string;
  /** Pre-filled LinkedIn people search URL */
  linkedinUrl: string;
  confidence: KeyContactConfidence;
}

/**
 * Whether the country has active hump yards
 */
export type CountryResultVerdict = typeof CountryResultVerdict[keyof typeof CountryResultVerdict];


export const CountryResultVerdict = {
  Yes: 'Yes',
  No: 'No',
  Uncertain: 'Uncertain',
} as const;

/**
 * Confidence level in the result
 */
export type CountryResultConfidence = typeof CountryResultConfidence[keyof typeof CountryResultConfidence];


export const CountryResultConfidence = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

/**
 * A=active modernization, B=active humps low spend, C=legacy base, D=no humps
 */
export type CountryResultTier = typeof CountryResultTier[keyof typeof CountryResultTier];


export const CountryResultTier = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface CountryResult {
  country: string;
  /** Whether the country has active hump yards */
  verdict: CountryResultVerdict;
  /** Confidence level in the result */
  confidence: CountryResultConfidence;
  /** A=active modernization, B=active humps low spend, C=legacy base, D=no humps */
  tier: CountryResultTier;
  /** One-paragraph intelligence summary */
  summary: string;
  /** Names of specific hump yards found */
  yards: string[];
  /**
     * Railway authority / operator name
     * @nullable
     */
  operator: string | null;
  /**
     * Last known modernization date or project
     * @nullable
     */
  lastModernization: string | null;
  /**
     * URL to procurement portal if found
     * @nullable
     */
  procurementPortal: string | null;
  /**
     * Named procurement contacts or railway authority directors
     * @nullable
     */
  contactEntryPoint: string | null;
  /** Recent procurement or modernization tenders found */
  procurementTenders: string[];
  /** Named technical decision-makers found */
  technicalContacts: string[];
  /** Structured key contacts with LinkedIn URLs */
  keyContacts: KeyContact[];
  sources: SourceLink[];
  /**
     * Error message if search partially failed
     * @nullable
     */
  error?: string | null;
}

export interface CountryList {
  countries: string[];
}

export interface OutreachInput {
  /** @nullable */
  contactName?: string | null;
  title: string;
  organisation: string;
  country: string;
  yards?: string[];
  /** Language code: 'en' (default), 'de', 'fr', 'ru', 'zh', 'pl', etc. */
  language?: string;
}

export interface OutreachResult {
  message: string;
  language: string;
}

export interface ErrorResponse {
  error: string;
}