# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hump Yard Market Intelligence API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: search
    description: Market intelligence search operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /search/country:
    post:
      operationId: searchCountry
      tags: [search]
      summary: Search a country for hump yard intelligence
      description: Uses Exa web search to find hump/classification yard data for a country
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CountrySearchInput"
      responses:
        "200":
          description: Country intelligence result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryResult"
        "500":
          description: Search error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /search/countries:
    get:
      operationId: getCountries
      tags: [search]
      summary: Get list of countries for Global Radar scan
      description: Returns the pre-loaded list of countries to scan in Global Radar mode
      responses:
        "200":
          description: Country list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryList"

  /search/outreach:
    post:
      operationId: generateOutreach
      tags: [search]
      summary: Generate a personalised outreach message for a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachInput"
      responses:
        "200":
          description: Generated outreach message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachResult"
        "500":
          description: Generation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CountrySearchInput:
      type: object
      required:
        - country
      properties:
        country:
          type: string
          description: Country name to search

    SourceLink:
      type: object
      required:
        - url
        - title
      properties:
        url:
          type: string
        title:
          type: string
        publishedDate:
          type: ["string", "null"]
        snippet:
          type: ["string", "null"]

    KeyContact:
      type: object
      required:
        - title
        - organisation
        - whyRelevant
        - linkedinUrl
        - confidence
      properties:
        name:
          type: ["string", "null"]
          description: Full name if found, null if only role is known
        title:
          type: string
          description: Role or job title
        organisation:
          type: string
          description: Railway authority or ministry
        whyRelevant:
          type: string
          description: One-line explanation of why this person matters for BD
        linkedinUrl:
          type: string
          description: Pre-filled LinkedIn people search URL
        confidence:
          type: string
          enum:
            - "Named & verified"
            - "Role known, name uncertain"
            - "Role inferred"

    CountryResult:
      type: object
      required:
        - country
        - verdict
        - confidence
        - tier
        - summary
        - sources
        - yards
        - operator
        - lastModernization
        - procurementPortal
        - contactEntryPoint
        - procurementTenders
        - technicalContacts
        - keyContacts
      properties:
        country:
          type: string
        verdict:
          type: string
          enum: [Yes, No, Uncertain]
          description: Whether the country has active hump yards
        confidence:
          type: string
          enum: [High, Medium, Low]
          description: Confidence level in the result
        tier:
          type: string
          enum: [A, B, C, D]
          description: "A=active modernization, B=active humps low spend, C=legacy base, D=no humps"
        summary:
          type: string
          description: One-paragraph intelligence summary
        yards:
          type: array
          items:
            type: string
          description: Names of specific hump yards found
        operator:
          type: ["string", "null"]
          description: Railway authority / operator name
        lastModernization:
          type: ["string", "null"]
          description: Last known modernization date or project
        procurementPortal:
          type: ["string", "null"]
          description: URL to procurement portal if found
        contactEntryPoint:
          type: ["string", "null"]
          description: Named procurement contacts or railway authority directors
        procurementTenders:
          type: array
          items:
            type: string
          description: Recent procurement or modernization tenders found
        technicalContacts:
          type: array
          items:
            type: string
          description: Named technical decision-makers found
        keyContacts:
          type: array
          items:
            $ref: "#/components/schemas/KeyContact"
          description: Structured key contacts with LinkedIn URLs
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceLink"
        error:
          type: ["string", "null"]
          description: Error message if search partially failed

    CountryList:
      type: object
      required:
        - countries
      properties:
        countries:
          type: array
          items:
            type: string

    OutreachInput:
      type: object
      required:
        - country
        - title
        - organisation
      properties:
        contactName:
          type: ["string", "null"]
        title:
          type: string
        organisation:
          type: string
        country:
          type: string
        yards:
          type: array
          items:
            type: string
        language:
          type: string
          description: "Language code: 'en' (default), 'de', 'fr', 'ru', 'zh', 'pl', etc."

    OutreachResult:
      type: object
      required:
        - message
        - language
      properties:
        message:
          type: string
        language:
          type: string

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string