# lib/api-zod/src/generated/types/countryResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hump Yard Market Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { CountryResultConfidence } from './countryResultConfidence';
import type { CountryResultTier } from './countryResultTier';
import type { CountryResultVerdict } from './countryResultVerdict';
import type { KeyContact } from './keyContact';
import type { SourceLink } from './sourceLink';

export interface CountryResult {
  country: string;
  /** Whether the country has active hump yards */
  verdict: CountryResultVerdict;
  /** Confidence level in the result */
  confidence: CountryResultConfidence;
  /** A=active modernization, B=active humps low spend, C=legacy base, D=no humps */
  tier: CountryResultTier;
  /** One-paragraph intelligence summary */
  summary: string;
  /** Names of specific hump yards found */
  yards: string[];
  /**
     * Railway authority / operator name
     * @nullable
     */
  operator: string | null;
  /**
     * Last known modernization date or project
     * @nullable
     */
  lastModernization: string | null;
  /**
     * URL to procurement portal if found
     * @nullable
     */
  procurementPortal: string | null;
  /**
     * Named procurement contacts or railway authority directors
     * @nullable
     */
  contactEntryPoint: string | null;
  /** Recent procurement or modernization tenders found */
  procurementTenders: string[];
  /** Named technical decision-makers found */
  technicalContacts: string[];
  /** Structured key contacts with LinkedIn URLs */
  keyContacts: KeyContact[];
  sources: SourceLink[];
  /**
     * Error message if search partially failed
     * @nullable
     */
  error?: string | null;
}