# lib/api-zod/src/generated/types/outreachResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hump Yard Market Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface OutreachResult {
  message: string;
  language: string;
}